requirements.txt
azure-functions
azure-eventgrid
azure-core
requests
